+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Change Log
 All notable changes to the project, past revisions, are documented in this file.
 
+### Version 0.0.9
+```
++ Added minor features to post new notices
++ Added announcements to database
+```
+
+### Version 0.0.8
+```
+Storage->Files
+  + Adapted to use ReactJS and JSX
+  + Able to change islocked attribute
+  + Able to add/remove user permissions
+  + Able to add/remove group permissions
+  - Permissions not enforced
+Admin->Permission (New)
+  + Able to view existing permissions
+```
+
 ### Version 0.0.7
 ```
 + API now requires login